Actually all tests is done

+++ src/test/java/LoginTest.java
@@ -5,23 +5,27 @@ import io.qameta.allure.junit4.DisplayName;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.html5.LocalStorage;
 import org.openqa.selenium.html5.WebStorage;
 import page_object.LoginPage;
 import page_object.MainPage;
 import page_object.RegisterPage;
 import page_object.ResetPasswordPage;
-public class LoginTest extends MainRule {
+public class LoginTest {
     MainPage mainPage;
     LoginPage loginPage;
     RegisterPage registerPage;
     ResetPasswordPage resetPasswordPage;
     private String accessToken;
     private User user;
+    WebDriver driver;
+    Browser browser = new Browser();
 
 
     @Before
     public void setUp() {
+        driver = browser.getWebDriver();
         mainPage = new MainPage(driver);
         loginPage = new LoginPage(driver);
         registerPage = new RegisterPage(driver);
@@ -83,7 +87,9 @@ public class LoginTest extends MainRule {
 
     @After
     public void cleanUp() {
-        UserSteps.deleteUser(accessToken);
+        if (accessToken != null) {
+            UserSteps.deleteUser(accessToken);
+        }
+        driver.quit();
     }
-
 }
